smurf: don't segv on un-initialized smfFilter
This can happen if you set flt.whiten. It fails miserably because
the filter itself maybe uninitialized and we're just trying to
call smf_whiten iteratively. Not sure of the best way to solve
this problem at the moment, but we should stay away from
flt.whiten... whitening can only be used as a pre-processing
step at the moment.