smurf: add ability to do common-mode removal at pre-processing
Just fiddled with smf_clean_smfArray so that it can call
smf_calcmodel_com. Currently set up to create its own COM/GAI
arrays that it uses locally and then drops on the floor. It also
makes its own copy of QUA. Need to get it to use previously
allocated arrays if available so that it can return the result
if we then choose to also fit for COM/GAI in the map-maker
iteratively.

Rather than having its own set of COM and GAI parameters, you
just set the "COMPREPROCESS" flag and it will use all of the
same COM.* and GAI.* parameters for the map-maker (which
required moving their defaults into smurf_sc2clean.def).