atl: Provide an option to allow objects to be read automatically from dumped Channels
If atl or kaplibs is used to read an AST Object from a text file, and the
file contains the dump of an AST Channel, then atl and kaplibs correctly
returns the Channel itself to the calling application, rather than the
Object contained within the Channel (if any). But sometimes it is useful
to have the Object within the Channel returned rather than the Channel
itself. So i've added a kludge which does this if the "Begin" line for
channel in the dump is of the form "Begin <class>(Read)", where <class>
is the channel class name. For instance, if the text file starts with

Begin FitsChan(Read)

the atl and kaplibs will return the object read from the FitsChan to the
calling app, rather than the FItsChan itself.