smurf: process data in smaller pieces for blank-field surveys
I set maxlen to 240 s (4 min.). This seems to result in slightly
deeper maps. I think this is just because the noise isn't stationary,
and we end up with a slightly more optimal weighting. 4 min. should
be enough time to complete several complete maps for all of the
blank-field SRO projects.