smurf: added ability to add extra astronomical sources to time series (fakemap)
By specifying "fakemap", the iterative map-maker will load in this
external file and add extra astronomical sources to the time series. The
fakemap must be double precision and have the same x- and y- pixel
dimensions as the map you are making -- this enables us to use the same
pointing LUT to sample off this map and add it to the time series. We
could certainly do something fancier in the future, but this is good
enough to get us up and running.

Also, note that you can do a fully simulated map by (i) running the
map-maker as usual, and (ii) creating a fakemap by *subtracting* the
real map off of your fake astronomical sources. In this way you will
remove the real astronomical signal, leaving only your simulated
signal (and noise...) in the time series.

Finally, if you are applying extinction correction, it will apply
the inverse of these factors to the fake signal as it is added
in to the time series so that the units work out in the end.