smurf: data cleaning needs to come after addition of fakemap data
Had to re-order things so that scanning the fakemap and adding
in the additional signal to the time series happens immediately
after creating RES, LUT and EXT, and then calling
smf_clean_smfData.

This also required allocation of the model pointers earlier
in smf_iteratemap since EXT was previously handled
after the allocation of the static models.