smurf: get inverse FFTs with SC2FFT working again
This was broken on a couple of levels.

First, we introduced the ability to concatenate input files before
doing the forward transformation. This doesn't make sense
for the inverse transformation, so to avoid special-casing a bunch
of the code I have reverted to single-file operation (you will
have to run sc2concat before forward transformations).

Also, we use smf_get_dims in a bunch of places to calculate things
like ntslices or ndata. This failed previously for 4d data, so
I have modified smf_get_dims to return sensible answers in
the 4d case using an internal call to smf_isfft.

Finally, smf_isfft can do a better job of calculating ntslices
if we have a smfHead attached. So, we check there first, and
then revert to the old method if necessary (checking that the
value at the Nyquist frequency is real/complex to decide if we
had an even or odd number of time slices).