smurf: add ability to calculate scan direction each timeslice
This will eventually be useful for producing sub-maps of data
scanned along certain position angles on the sky (and hopefully
lead to an improved method of adding together the data to preserve
larger-scale information).

The best place to do the calculation seemed to be smf_coords_lut
which currently works out the pointing lookup table for the output
map. I've extended smfData with a "theta" pointer that gives the
scan direction at each time slice (along with required initialization
and free in smf_create_smfData and smf_close_file respectively).