smurf: Change int jcmtstate entries to shorts
To save a tiny bit of memory convert all the integer entries
in JCMTSTATE to shorts in the struct and in the file. Old data
will be converted on read although conversion errors are
assumed to be irrelevant since all except bad values will be
in range.

The SCUBA-2 bias and heater entries are unsigned shorts.