KAPPA: Divide the processing according to line-fit subtraction flag.
As currently invoked this routine can try to read from the data array
mapped with write access in MFITTREND.  So it only tests for supplied
values if subtracting the fit from the data, instead of storing the
fit in the data array.  While MFITTREND could initialise the fit array
to 0, it's a fudge.  The extra code added has the further efficiency
advantage that the subtraction flag is tested only once instead of
checking every element, and there are no longer unnecessary bad-pixel
tests during storage of the evaluated line fit.