smurf: Add JCMTSTATE compression
This patch enables data compression of JCMTSTATE. A new routine,
sc2store_writejcmtstate, is added to sc2store to create the
JCMTSTATE extension with all its components. This is required
because we can not create the components first and then fill
them without knowing how big they can be.

The compression is very simple. We just check to see if the value
is a constant for the entire file and if so we write a single
value as a scalar. The read routines have been modified to expand
scalars into vectors of the relevant size.

This code takes about six times longer than the previous
implementation but is still fast. 6000 state entries can be written
in about 0.035 sec on my Mac. A test file I was using shrinks
by about 1%.