smurf: Rewrite grouping algorithm
The old algorithm relied on the input files being sorted into
time order and on the values of rts_end. The new version fills
an AST keymap with values that will be unique for a group of
related files and then ensures that the keymap is read out
in date order.