ndg: Allow prov structures to be created that are not associated with an NDF
The ndgReadProv function can now be supplied with an NDF identifier of
NDF__NOID, in which case an empty provenance structure is created and
returned. Ancestors can be added to this empty provenance structure as
usual. When the provenance structure is written out to an NDF, the path
to the NDF will be added to the resulting provenance extension (but not
to the in-memory provenance structure). Thus if an app creates many
output NDFs from a single set of input NDFs, it can create an empty
provenance structure, add all the input NDFs to it, and then write the
same output provenance structure out to all the output NDFs without the
need to reconstruct it from scratch for each output NDF.