smurf: Change flagslow for pointing dimmconfig
Observation 20091208#55 does not reduce with the default
pointing config because of an interaction between the
new filter code and the fact that 60% of the data are flagged
as being too slow. Lowering flagslow to 20 fixes the problem
but this may not be the real fix as the data reduce fine
using the bright_compact config.