convert: make ndf2fits write out wcs for each array component
This reverts the change to cof_whead included in bf752fe0. Each
IMAGE extension needs to have its own WCS so that  it can be
handled independently of the other IMAGE extensions. (Also, the
bf752fe0 change was bad as it resulted in WCS inherited from
the NDF's FITS extension being stored instead of the potentially
modifed WCS from the NDF's WCS component).