smurf: calculate correct units for power spectral densities
When we set "power" in sc2fft we really want to get a power
spectral density (this is what people usually mean by a power
spectrum). So, we divide each squared fourier amplitude by df --
the frequency spacing of each bin. In this way the amplitude of
the noise won't, for example, depend on how long the time series
was that you are analyzing.

I have noticed that inverse fft from this PSD representation
seems to be broken (sc2fft power=true inverse=true), though
unrelated to this patch.