smurf: update identification of outliers in map-based despiker
Previously there was a sort of average scale factor worked out
to normalize weighted samples so that they could be compared to
the estimated population variance (weightnorm). I've now realized
that a better local estimate can be calculated using the
map weights array. I've also updated comments in a few places
to clarify what is going on. I've noticed that this tends to
flag fewer samples than before (~30% or so), so you may notice
minor differences. Might consider dropping the threshold
ast.mapspike.