ndg: Speed up reading and writing of large provenance systems
As of NDG V6 provenance info in an NDF "PROVENANCE" extension is stored
in a single array of integers, rather than in lots of separate HDS
components. This is much faster to access and occupies far less space.
However, it means that new data created with V6 provenance will not be
readable by applications linked against earlier versions of NDG. But of
course NDG V6 can read either type of provenance.