smurf: I broke down and started adding PCA cleaning to the map-maker
This is a draft of the routine for doing PCA cleaning. It presently
only calculates the projection of the bolometers without actually
doing the filtering (the easy part). As feared it is very slow:
dominated by operations like calculating the bolo-bolo covariance
matrix (needs to take the dot product of all unique pairs of
bolometers). Threading will help, but this will always be a
painfully slow routine.

It's already bad enough with 1 subarray that I'm not planning
on trying to get it to do multiple subarrays simultaneously.