smurf: add new task SC2PCA to SMURF
This is a wrapper for the principal component analysis routine that
does the projection along the statistically independent components.
This routine does not do any actual filtering, instead this
capability will be added at a future date to SC2CLEAN. The purpose
of this new task is to make it easy to look at correlated noise
components in a data file that has already been pre-processed with
SC2CLEAN. The resulting components data cube has a full header so
that it can be run through sc2fft with power=true to look for
line features in the correlated noise.

Also, this routine does not concatenate. Since this is mostly for
exploring noise properties, I assume that if you throw a bunch of
files at it you are interested in seeing the PCA done on each
individually, rather than the full data set (realizing of course
that this routine is REALLY slow, despite having parallelization).