smurf: Change distortion script to expect bolomap names to include chunk number
When the distortion script was written, makemap omitted the leading chunk
number in the bolomap name if only one chunk was used. Now, makemap
always includes the leading chunk number, so change the distortion script
to always include a chunk number in the bolomap name.

ast: Improve distribution of points produced by astGetRegionMesh