smurf: smf_clea_pca can now actually clean...
Added the last bit to this routine. For each component, the
RMS of the amplitudes for all bolometers is calculated -- a positive
number giving the average amplitude for a given component. An
iterative sigma-clip using the previously unused "thresh" parameter
decides which of the large outlier modes should be removed. On
a random bit of test data, a value of thresh=4 seems to be good.
The supplied smfData are cleaned in-place if thresh is nonzero.