smurf: PCA cleaning is now a pre-processing option
Ensure that order=0, and then try setting pcathresh=4, either
in a call to sc2clean, or makemap. If you decide to run makemap,
you might as well just use PCA and forget COM and FLT, i.e.:

^/stardev/share/smurf/dimmconfig.lis
numiter = 2
pcathresh=4
modelorder = (ext,ast,noi)

Where the reason for 2 iterations is so that the noise weights
calculated after the 1st pass are used to make the map.