smurf: free up extra quality array in RES
Once the QUA has been initialized with the quality stored in
RES that resulted from concatenation, we can free it up because
the contents of QUA are used universally as a sidecar component.

This should free up a bit of memory, and make it more obvious
if there are any places that we missed proper implementation
of sidecar quality.