smurf: modifications to PCA cleaning to allow bad-bolo rejection
The first component of the PCA should be just as good at identifying
bad bolometers as the common-mode. In order to use smf_find_gains
we therefore need equivalent parameters -- for now we just use
the com.* ones, although in the future (probably when a PCA model
is added) we will make a set of pca.* versions. The other interesting
problem I ran into is the arbitrary sign of the components that
comes out of the PCA. smf_find_gains throws out bolometers that have
a negative gain WRT the template (to catch gross flatfield errors
primarily). So, I now check the average values of the amplitudes
from the PCA, and flip the sign if on average they are negative.