smurf: Allow user to specify MAKECUBE reference pixel coords
The new output parameter PIXREF records the interim grid coords at the
reference point. The new input parameters REFPIX1 and REFPIX2 specify the
interim grid coords to be placed at the reference point (if supplied,
they override the defaults supplied by the autogrid algorithm).

This allows the projection determined by the autogrid algorithm to be
re-used on subsequent runs of makecube.