smurf: don't chunk bolomaps
Previously the bolomaps from each data chunk would be written to
a different extension. This behaviour was not particularly useful,
so the contchunk parameter to smf_write_bolomap has been dropped,
and each time a new chunk is processed it will be added to previously
regridded data for a given subarray, column and row.