smurf: common-mode can now be used just for flagging
Added a parameter to the config file called com.noremove which
ensures that the common-mode is not actually removed from the
time-series. This is useful because smf_calcmodel_com is quite
good at identifying bad data, and in some cases it is still
nice to leave the low-frequency correlated noise in for the
remaining good data.