smurf: New distortion script - distmap
The distmap script is based on the bolomaps method from the old
"distortion" script. It is re-structured to separate the creation of the
bolomaps from the analysis of the bolomaps. This means that new
observations can be included in the map without the need to re-process
all the old data each time.

The old "distortion" script is retained because it has other facilities
not provided by the new script.