smurf: New way of synchronising related jobs
When calling smf_add_job, you used to be able to specify a function that
could be used to determine if a job was ready to run. This is clumsy and
awkward to use (and in fact was not being used anywhere). So I've changed
it so that you now just specify a list of identifiers for jobs that must
complete before the new job can run. This is paving the way for threading
the file opening and flatfielding code.