smurf: fix logic in choosing extinction monitor
It looks like we should be checking the time at the start AND end.
There was a case, https://forge.roe.ac.uk/trac/SCUBA2/ticket/464,
in which the start of the WVM time-series was junk, but the end
was OK. Under the old logic it would still decide to use WVM and
bad values would creep in.