smurf: Use correct DTOI value
We have been using an incorrect value for D/A to current conversion
because by the time we realised the error we did not want to confuse
the calibration for SRO.

Now we are about to update the resistor values we can also fix this
conversion.

Note that "burned in" flatfields for SCUBA-2 data before 20100223
will have to be recalculated and the FCFs for SRO data will also
have to be modified.