smurf: first working version of TMP generic template fitting model
This model has been set up so that it is relatively easy to fit
templates based on things like the JCMTState time-series to
each bolometers. The model container is the same dimensions
as GAI (3 planes: gain, offset, correlation coefficient), although
currently I just hard-wire it to remove the mean from each template,
and only a gain is fit. No multi-threading yet but it's pretty
fast. Some tests on az scans in the dark suggest that it works
very well.