pcs: Add malloc sanity checks for subpar.c
These were mainly driven by my attempts to replace
starmem malloc with something that would fail occasionally.

The main thing to learn from this patch is that all the
CNF string macros are a bit flaky if any of the CNF string buffers
fails to get malloced. We then pass these pointers straight into
the Fortran layer to get an immediate SEGV.