smurf: implement raw2current conversion that depends on date
We used to have RAW2CURRENT as a macro in smurf_par.h. However.
there was just an MCE firmware update to accomodate faster sample
rates and the low-pass filter changed, resulting in a new scale
factor. This scale factor is now determined from the date of the
observation in smf_raw2current().