smurf: Add new command unmakemap
This is like unmakecube but for SCUBA-2 rather than ACSIS - it creates a
set of time series cubes from a 2D map of the sky and set of template
time series cubes. It can probably be tuned to make it more generally
useful, but at the moment it does what I need. My motivation was
testing the distortion script (i.e. use unmakemap to generate a set of
time series with known distortion, and then see if the distortion script
can recover the known distortion).