smurf: continue gracefully for very large/low filter frequencies
Previously bad status would be generated in smf_get_findex if,
for example, you tried to low-pass filter below 30 Hz, but
you have 20 Hz down-sampled data (since the requested filter
edge is greater than Nyquist). Most of the time the most useful
behaviour is for smf_get_findex to simply return the Nyquist
frequency index if you happen to request a frequency that is
higher than Nyquist. Similarly, an error was generated if
the requested frequency would be rounded to 0 Hz. Again, this
is normally OK.