ndg: Fix problem with provenance hash codes used to identify NDFs
kappa:paste was crashing with a seg fault because the provenance system
had produced a situation where an NDF was its own ancestor, thus causing
an infinite recursion loop. This turned out to be due to two different
NDFs being assigned the same hash code. Changign the hash code algorithm
from 32 bit to 64 bit did not fix this. There seems to be some
deterministic reason why these two NDFs end up with the same hash code,
but I can't work out what it is. A work-around seems to be to add an
extra shift of one bit to the hash code when combining the hash code of a
parent with all the hash codes of its children.