ndg: Equality of hash codes does not guarantee that the objects are equal
Duh! Pretty obvious. The provenance system previously assumed that if two
hash codes were equal then the NDFs were equal. Now, if the two hash
codes are equal, it goes on to do an explicit check of the properties of
the two NDFs. This would have been a better way of fixing yesterdays
provenance problem.