smurf: add optional median calculation to smf_stats1
There is a comment in smf_mask_noisy that mentions there is no
decent thread-safe median routine available to us. I added this
functionality to smf_stats1 where it (optionally) uses the
system qsort (stdlib.h) to sort the data. I think this will be
useful for flagging outlier noisy bolometers; the NEP arrays
are fairly small.

Note that since we have the array sorted it would also be
trivial to add other interval-like quantities. For example,
instead of reading off the central value to get the median
(index/2), we could get the 1st and last quartile statistics
simply by reading off index/4 and index/3.