smurf: add median capability to smf_clipped_stats1D
Since this is a wrapper for smf_stats1 it is easy to also
calculate the median using the new functionality. It is
then possible, with the usemedian parameter, to clip
N-sigma outliers from the median, rather than the mean.
This ought to be faster / safer at removing the tail from
a skewed distribution.