smurf: add ability to clip NEP and NOISE values in CALCNOISE
Things like weighted average NEPs and effective NEPs were getting
thrown-off if the responsivity masking failed to catch a bolometer
that went dead some time after a setup. It would then exhibit very
small fluctuations in its time-series, and have an incorrect (high)
responsivity that would make it look like an extremely low-noise
detector (and therefore massively throw-off 1/noise^2 type
weightings). The defaults are now setup to throw out low-outliers
based on the scatter in the log(NEP), using the median instead of
the mean as the central measure. Seems to work well on a few
examples.