smurf: heavier clipping for sc2tsnoise
Some subarrays, in particular s8d, would still exhibit a large
population of bolos that would go dead a while after the
dark flatfield. This population lies fairly close to the peak
of the distribution, hence the heavier clipping on the low-side.
Also decided to throw out the really large positive outliers.