smurf: new downsampling method based on truncating FFT
I was wondering why downsampled data seem to produce noisier maps
(well, sometimes), and I thought that aliasing of the high-frequency
noise using the old time-domain boxcar smoothing method might be
an issue. It turns out that the new method isn't any better. Still,
I may experiment with it in the future. I also messed with a number
of small things:

- you can specify downsampfreq instead of downsampscale if you
  have a particular sample rate that you want

- the low-level smf_dataOrder_array routine can now convert
  SMF__INTEGER to SMF__DOUBLE in addition to re-ordering, which
  adds extra functionality to smf_deepcopy_smfData

- smf_fft_data can now also take SMF__INTEGER as an input type
  do to the above change. I also fixed up some dimensionality
  bugs/omissions

- the smf_get_padding and smf_grp_related routines don't need to
  have downsampling parameters in their interfaces because that
  information is already in the keymap