smurf: Allow sc2ast to read distortion from a PolyMap dump in a file
The SMURF_DISTORTION environment variable can now be of the form
"CUSTOM:<file>", where <file> is the path to a text file holding a dump
of a PolyMap followed by the dump of a ZoomMap. The PolyMap should
describe the required polynomial distortion, and the ZoomMap should
implement the conversion from mm to radians.