smurf: both calcnoise and smf_mask_noisy share outlier clipping routine
The purpose of this is to enable low- and high- outlier rejection,
although some additional config parameters will be needed. We can
now also use the log instead of the noise values themselves. Finally,
this new routine uses the median as a central measure, so this ought
to be more robust.

In this commit the behaviour shouldn't be too different from the
old behaviour because the log of the values is turned off, although
the median is used instead of the mean.