smurf: version 1.3.3, high- and low-outlier noisey bolo clipping
Finished implementing the clipping of both high- and low-noise
bolometers. We now use the median and standard deviation of
the log of the noise distribution, instead of the mean and
standard deviation of the noise distribution.

Instead of noiseclip we now have two parameters called
noisecliphigh and noisecliplow (the latter of which defaults
to 0 to turn it off).

Since this modifies the syntact for config files slightly
I've bumped the version number of SMURF up a notch.