smurf: add SAMPCUBES optional extension to makemap
This was requested in ticket #726 by Antonio. For each chunk of
data, by setting the sampcube=1 flag in the confuguration file,
an extension can be created containing a cube with the map
dimensions along one face, and samples that go into each map
pixel along the third dimension. A VARIANCE component is also
generated, so it is possible to reproduce the same weights
used when rebinning the data internally (1/VARIANCE).