smurf: add smf_fix_data to handle may/june 2011 s4a row readout error
smf_fix_data is based on smf_fix_metadata, and is called at the end
of smf_open_file. This routine only does a date / subarray number check
to decide whether the row fix is needed or not. It would be better for
this routine to actually check the MCE header for inconsistencies, but
this seems to be good enough for now. As long as fastflat data are
included when running makemap you should get a consistent data file /
flatfield set.

The hard-wired date range that is checked is (in UT)
2011-05-26T19:00:00 to 2011-07-06T10:23:00