TWODSPEC: Fix workspace handling in line fitting.
There was some further offsets done manually instead of using
DYN_INCAD as needed for 64-bit addressing.

For the first tranche of workspace replaced a GETVM call with straight
DSA_GET_WORKSPACE so as not to undermine DYN_INCAD. Added some
comments to make it clearer where the first tranche is tidied.

For the second tranche the division via multiple pointers is clarified
using DSA calls rather than DYN_INCAD, and this makes tidying easier,
and freed this space.