smurf: Use better WVM225 to CSO225 conversion factor
The previous conversion was very bad in that it actually had a typo
that made it go the wrong way. This made things very bad. This new
value comes from Remo Tilanus. The conversion is now in a separate
routine so that it can also be applied to the WVM FITS header.