smurf: flag to disable COM bolo flagging, and export COM/GAI with SC2CLEAN
We didn't have a way of turning off bad-bolo flagging when we remove the
common mode, so I added a COM.NOFLAG option. This is to make it
easier to test the effects of using common-mode cleaning vs. polynomial
baseline removal on CALCNOISE for recent data that would often flag
far too many bolometers. However, the default in dimmconfig_calcnoise.lis
is still to use polynomials, although I fleshed out the comments.

We can now also write out the COM and GAI models when we run sc2clean
(the smfArrays are returned in smf_clean_smfArray so other routines
can now also access these data if we care, like CALCNOISE).

Finally, small tweak to smf_write_smfData so that you can specify the
message level at which the filename being written is reported.