smurf: use noiseclipprecom to flag noisy bolos before common-mode removal
Common-mode removal depends on the removal of bad bolometers. Normally
we have the COM model iteratively throw-out bolos that don't match
the common-mode template, but this can be problematic for noise
measurements which (a) may have very short time series (so hard to
build up a template with structure that can be fit), and (b) we
want to keep as many bolos as possible for the purposes of fitting.
It seems that a reasonable approach is to optionally flag/removal
the noisiest bolometers immediately before common-mode removal
using this new parameter "noiseclipprecom". In the future, a nicer
approach would be to have a completely dynamic order for the
cleaning options, much like modelorder for the map-maker.