smurf: add dimmconfig_calcnoise_com.lis and dimmconfig_calcnoise_pca.lis
I re-jigged the calcnoise configuration files so that we now have the
default (dimmconfig_calcnoise.lis) that just uses polynomial baseline
subtraction (robust, and fast, but perhaps noisy and biased upwards),
as well as a variant that uses the new common-mode removal functionality
(dimmconfig_calcnoise_com.lis) and a PCA version which is extremely
slow (dimmconfig_calcnoise_pca.lis). My thought is that the default
should be used in the pipeline, and probably the other two should
only be used for research purposes (PCA is probably best, though
not tested on a variety of data yet).