surf: Add SKYDIP2 task which can fit instrument agnostic skydip data
Takes a simple NDF as input containing sky temperatures and airmass.
Assuming a few FITS keywords are present this can be used to fit
a Skydip using the standard SCUBA skydip model. The aim is to
be able to use this command to reduce SCUBA-2 skydip data.

Can be tested using the output from the SKYDIP command.