kaplibs: Ensure test point returned by kpg1_asgdp is not close to the origin
Using the origin as a test point is a bad idea as it has unusual
properties (e.g. you cannot change the value by scaling it). So
kpg1_asgdp now ensures that the returned point is more than 10% of the
box size away from the origin in both input and output coordinates.

This solves the problem reported by Jess in which kappa:collapse issued
the error: "The specified WCS axis (axis 2) has a constant value over the
whole NDF and so cannot be collapsed." when in fact WCS axis 2 did not
have a constant value.