smurf: fix combination of maps from each chunk. Bump SMURF to 1.3.6
After some good sleuthing by DSB, and also studying the noise properties
of our Lockman Hole map from S2SRO, I discovered a bug that has been around
a long time in regards to how maps made from each disjoint chunk of
data are added to the sum of the previous maps -- like what CCDPACK:MAKEMOS does
(indeed, Andy mentioned a long time ago that MAKEMOS, as used in the pipeline,
was getting slightly different values for the noise than MAKEMAP).

The problem was this: when we create maps with smf_rebinmap1, the usual
behaviour is that we use weights for each data point calculated from
the white-noise level in the bolometer power spectra. However, weights
can in general be anything you want, and in practice this estimate of
the noise is off by some factor due to filtering, the fact that the bolometers
do not have perfect white noise etc. For that reason, smf_rebinmap1 calculates
the variance in a map pixel using the (weighted) scatter of the data themselves
rather than blindly propagating the estimated variances. This is all fine
if we are only making a map out of a single continuous chunk of data. However,
smf_iteratemap uses smf_addmap1 to combine the sub-maps from different chunks.
What I *should* have been doing is the 1/variance-weighted average of the
new map with the accumulated previous maps, i.e.

             totalmap = (oldmap/oldvar + newmap/newvar) / (1/oldvar + 1/newvar)

    and      totalvar = 1 / (1/oldvar + 1/newvar)

Instead, I was using the accumulated *weights* for oldmap and newmap
when doing the average to calculate both the total map and total
variance (i.e., assuming that weight = 1/variance). However, as I
described before, the weights are, under pretty much the best
circumstances, only close to being 1/variance.

Since the weights are probably OK in a relative sense, the maps
themselves hardly changed, but the variance maps could be wrong
by quite a bit -- especially if there was a lot of filtering
in the time-series beforehand (i.e. the stuff Antonio has been
doing lately with heavy low-pass filtering).

For the most part, pipeline-reduced data will be OK provided that
the machine it was run on had enough memory to avoid breaking
single maps into multiple pieces. It then uses MAKEMOS which
does the weighted average of the sub-maps correctly.

This can have a pretty major impact on large data sets reduced only
with SMURF though, hence the increase in the version number.