kappa: Change setaxis so that it works for crossed WCS axes
Previously "setaxis mode=wcs dim=2" assumed that the AXIS(2) structure
should be filled with values from the second WCS axis. But if the WCS
axes are crossed (or not parallel to the axes) it may be better to use some
other WCS axis. SETAXIS now tries to find the best WCS axis to use.