smurf: Fix heater reistance and apply heater efficiency data
The reference heater resistance is now fixed at 3ohms and the
per-bolometer resistance is now fixed at the reference resistor
value and is no longer specified in the resist.cfg file.
smf_flat_standardpow therefore still runs but will never regrid
the heater data (so TABLE mode still works).

Heater efficiency files have been created based on the current
best resistor values but now referenced to 3 ohms. Responsivity
images are now corrected by these values inside
smf_flat_responsivity.

The heater efficiencies are applied as part of flatfielding
rather than being applied when the flatfield itself is
calibrated. This allows us to shift the calibration to
the off-line system without having to worry about new
reistance values forcing all previous flatfields to be
recalculated.

Now smf_find_science looks at all the ARRAYID headers and
caches the corresponding heater efficiency data in an AST
keymap. This keymap is then passed to every routine that
previously took flatramps so much of this commit is simply
changing associated code to take an extra argument. I
decided not to read the efficiency data when a file
was opened (to populate smfDA immediately). I also decided
not to read the efficiency file on demand inside smf_flatten.
The major reasoning being that I do not want to open the
same efficiency file every time an s8a file is opened. There
is also the issue that we now run flatfielding inside a thread
so can not call NDF.

Flatfielded data do now have the same output values and
responsivities as they had before this commit but the
heater efficiencies are easier to manage.

Note that as things stand the heater efficiency will be
applied even if the flatfield was previously calculated
with a resistor value not equal to 3 ohms. This is a
serious issue and the fix requires us to modify
sc2store to write the resistor value (as a surrogate
for a version number). If we can read a resistor value
from the flatfielding information we know we have the
new flatfielding code and can apply the efficiency.

The other thing missing is the ability to write out the
efficiency data that was applied to the output time
series.