smurf: fix broken maxlen multiple subarrays and/or down-sampling
The "maxlen" parameter allows you to specify a maximum length
in time that you want to concatenate. This feature was broken
in two ways: (1) it was calculated before down-sampled steptime
was estimated, so it would give the wrong value if downsampling
was turned on, (2) it was being calculated multiple times, and
getting different answers, giving odd results primarily when
multiple subarrays were used. smf_grp_related has been patched
to have sensible behaviour.