smurf: Do not apply heater efficiencies if we are using older flats
All new flatfield calculations store the reference resistor value
in the flatfield extension of data files and in the new refres
field of smfDA. Older flatfields will not have that value in
smfDA so the system assumes they do not need to have the
heater efficiency applied (it usually means they were calculated
by modifying the heater resistance directly).

This has required changes to the sc2store API to read and write
refres so quite a few fles have been tweaked with this patch.