smurf: Add sky offset back on to data before fitting flat poly
POLYNOMIAL flatfields were correcting for the DC offset when
they were applied. Fitting the sky background and removing
it meant that the POLYNOMIAL flatfield solution did not
have access to the sky numbers. This was preventing cubics
from ever working since the reference sky level was always
being treated as 0.

Now we add back the value that was calculated from the polynomial
sky fit as seen at the end of the observation.