smurf: Fix sky fitting for fast flats
The extrapolated value at the reference heater setting was
not being included properly in the sky fit. The DAC units
value was correct but the code assumed we were fitting DAC units
as a function of heater whereas we were fitting DAC units
as a function of index into the data array. Needed to calculate
the indices for the relevant sky value.

The effect of this was to force the fit to go through a point
that was far removed from the data in x.