smurf: add isFFT to smfData and record in FITS header
First step toward adding 2D FFT capabilities is to clean up the
way we store whether we have FFT data or not. I have added a
flag called isFFT to the smfData, and a FITS header with this
value is also created whenever we do a forward/inverse transform
with smf_fft_data. For now I have set it up so that 0 is the
default value (we don't know) and the smfData constructor sets this
value. When we open the file we read in the value of the keyword.
If we have a data file with a FITS header but no ISFFT keyword (like
raw data) it is safe to assume that this is not the FFT of something.
For data that are not FFT we set it to -1. Finally, if this is the
FFT of something, we set the value to the length of the last
real-space axis. The reason for this is that when we do a forward
transform (and since we are only ever working with real-valued data),
due to symmetries FFTW doesn't bother storing half of the redundant data,
and the last axis is only n/2+1 elements long. However, we need to know
whether n is even or odd in order to get the dimensions right for
the inverse transform... so for now we'll store it here.