smurf: re-vamp smf_isfft to use smfData.isFFT flag
The guess work based on the real/imaginary parts of the last
dimension have been removed since we now store that length
in the smfData.isFFT flag. This function still returns a boolean
even though smfData.isFFT can be -1 (real-space data),
0 (ambiguous), or +nsamples (if FFT data). However, it not does
a bunch of internal checks to decide whether things are consistent
or not. If not, or smfData.isFFT is 0, it can't figure out the
data and sets bad status. Since the smfData constructor defaults
to 0, this means that calling this routine will set bad status,
helping to catch cases where you have not either (a) loaded in the
isFFT flag from the FITS header of an existing data file, or (b)
explicitly set it for the particular situation.

The other big difference is that it has now been extended to handle
2D data in preparation for FFTs of maps.